+++ commitlint.config.js
@@ -10,9 +10,9 @@ module.exports = {
       scopes.rules['scope-enum'](ctx).then(([level, applicable, packages]) => [
         level,
         applicable,
-        packages.map(
-          pkg => (pkg.includes(prefix) ? pkg.replace(prefix, '') : pkg)
-        )
+        packages
+          .map(pkg => (pkg.includes(prefix) ? pkg.replace(prefix, '') : pkg))
+          .concat(['examples'])
       ])
   }
 }