+++ packages/halo/stories/index.js
@@ -4,21 +4,11 @@ import { storiesOf } from '@storybook/react'
 import Halo from '../react'
 import Focusable from './focusable'
 
-const appearanceStories = storiesOf('Halo/appearances', module)
-
-Object.keys(Halo.appearances).forEach(appearance => {
-  appearanceStories.add(appearance, _ => (
-    <Halo appearance={appearance}>
-      <Focusable>{appearance}</Focusable>
-    </Halo>
-  ))
-})
-
 const gapSizeStories = storiesOf('Halo/gapSizes', module)
 
 Object.keys(Halo.gapSizes).forEach(size => {
   gapSizeStories.add(size, _ => (
-    <Halo gapSize={size}>
+    <Halo gapSize={size} visible>
       <Focusable>{size}</Focusable>
     </Halo>
   ))
@@ -28,7 +18,7 @@ const shapeStories = storiesOf('Halo/shapes', module)
 
 Object.keys(Halo.shapes).forEach(shape => {
   shapeStories.add(shape, _ => (
-    <Halo shape={shape}>
+    <Halo shape={shape} visible>
       <Focusable shape={shape}>{shape}</Focusable>
     </Halo>
   ))
@@ -42,12 +32,45 @@ storiesOf('Halo/visible', module)
   ))
   .add('false', _ => (
     <Halo visible={false}>
-      <Focusable>not visible</Focusable>
+      <Focusable>not visible (until focused)</Focusable>
+    </Halo>
+  ))
+  .add('false & not visibleOnFocus', _ => (
+    <Halo visible={false} visibleOnFocus={false}>
+      <Focusable>really not visible (even on focus)</Focusable>
+    </Halo>
+  ))
+  .add('true & error', _ => (
+    <Halo visible error>
+      <Focusable>visible & error</Focusable>
+    </Halo>
+  ))
+  .add('false & error', _ => (
+    <Halo visible={false} error>
+      <Focusable>not visible (always visible if error) & error</Focusable>
     </Halo>
   ))
 
-storiesOf('Halo/visibleOnFocus', module).add('enabled', _ => (
-  <Halo visible={false} visibleOnFocus>
-    <Focusable>visible when focused</Focusable>
-  </Halo>
-))
+storiesOf('Halo/visibleOnFocus', module)
+  .add('enabled', _ => (
+    <Halo visible={false} visibleOnFocus>
+      <Focusable>visible when focused</Focusable>
+    </Halo>
+  ))
+  .add('enabled & error', _ => (
+    <Halo visible={false} visibleOnFocus error>
+      <Focusable>visible when focused with error</Focusable>
+    </Halo>
+  ))
+
+storiesOf('Halo/error', module)
+  .add('true', _ => (
+    <Halo error>
+      <Focusable>error</Focusable>
+    </Halo>
+  ))
+  .add('false', _ => (
+    <Halo>
+      <Focusable>not error</Focusable>
+    </Halo>
+  ))