+++ packages/button/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.8.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-button@10.8.17...@pluralsight/ps-design-system-button@10.8.18) (2018-11-09)
+
+
+### Bug Fixes
+
+* upgrade storybook and fix issues with require.context ([ba6a6e8](https://github.com/pluralsight/design-system/commit/ba6a6e8))
+
+
+
+
+
 ## [10.8.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-button@10.8.16...@pluralsight/ps-design-system-button@10.8.17) (2018-10-31)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-button