+++ packages/card/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.5.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-card@8.5.3...@pluralsight/ps-design-system-card@8.5.4) (2018-11-09)
+
+
+### Bug Fixes
+
+* upgrade storybook and fix issues with require.context ([ba6a6e8](https://github.com/pluralsight/design-system/commit/ba6a6e8))
+
+
+
+
+
 ## [8.5.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-card@8.5.2...@pluralsight/ps-design-system-card@8.5.3) (2018-10-31)