+++ packages/datepicker/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datepicker@0.4.15...@pluralsight/ps-design-system-datepicker@0.4.16) (2018-11-09)
+
+
+### Bug Fixes
+
+* upgrade storybook and fix issues with require.context ([ba6a6e8](https://github.com/pluralsight/design-system/commit/ba6a6e8))
+* **datepicker:** update/parse new calendar values when state changes ([bf6244f](https://github.com/pluralsight/design-system/commit/bf6244f)), closes [#304](https://github.com/pluralsight/design-system/issues/304)
+
+
+
+
+
 ## [0.4.15](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-datepicker@0.4.14...@pluralsight/ps-design-system-datepicker@0.4.15) (2018-10-31)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-datepicker