+++ packages/icon/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.13.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-icon@8.13.2...@pluralsight/ps-design-system-icon@8.13.3) (2018-11-09)
+
+
+### Bug Fixes
+
+* upgrade storybook and fix issues with require.context ([ba6a6e8](https://github.com/pluralsight/design-system/commit/ba6a6e8))
+
+
+
+
+
 ## [8.13.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-icon@8.13.1...@pluralsight/ps-design-system-icon@8.13.2) (2018-10-31)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-icon