+++ packages/link/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.7.15](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-link@5.7.14...@pluralsight/ps-design-system-link@5.7.15) (2018-11-09)
+
+
+### Bug Fixes
+
+* upgrade storybook and fix issues with require.context ([ba6a6e8](https://github.com/pluralsight/design-system/commit/ba6a6e8))
+
+
+
+
+
 ## [5.7.14](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-link@5.7.13...@pluralsight/ps-design-system-link@5.7.14) (2018-10-31)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-link