+++ packages/storybook-addon-center/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-center@1.0.2...@pluralsight/ps-design-system-storybook-addon-center@1.0.3) (2018-11-09)
+
+
+### Bug Fixes
+
+* upgrade storybook and fix issues with require.context ([ba6a6e8](https://github.com/pluralsight/design-system/commit/ba6a6e8))
+
+
+
+
+
 ## [1.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-center@1.0.1...@pluralsight/ps-design-system-storybook-addon-center@1.0.2) (2018-10-31)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-center